+++ src/App.tsx
@@ -7,6 +7,7 @@ import Calendar from './pages/manage/calendar';
 
 const ManagePatients = lazy(() => import('./pages/manage/managePatients'));
 const Settings = lazy(() => import('./pages/manage/settings'));
+const TestComponents = lazy(() => import('./pages/testComponents'));
 
 function App() {
   return (
@@ -16,6 +17,7 @@ function App() {
           <Switch>
             <Route exact path={'/patients'} component={ManagePatients} />
             <Route exact path={'/settings'} component={Settings} />
+            <Route exact path={'/components'} component={TestComponents} />
             <Route exact path={'/'} component={Calendar} />
             <Route path={'*'} component={Notfound} />
           </Switch>