+++ src/pages/manage/managePatients.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 const ManagePatients = () => {
-  return <div>manage patients</div>;
+  return <div>Manage Patients</div>;
 };
 
 export default ManagePatients;