+++ CHANGELOG.md
@@ -2,6 +2,36 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.2.0](https://github.com/giftdavid101/appointment-schedular/compare/v1.1.0...v1.2.0) (2021-10-03)
+
+### Features
+
+- add truncate props ([d930379](https://github.com/giftdavid101/appointment-schedular/commitsd930379a906f6c652b8576e2afa6c62ca2fc3918))
+- add helper function to convert Hex color to HSL ([2d4faa3](https://github.com/giftdavid101/appointment-schedular/commits2d4faa32caa492df9d0d205e857a1d36c3e5a08b))
+- add test component route for development ([73df2f7](https://github.com/giftdavid101/appointment-schedular/commits73df2f7e734f20bd5a027aed56f44e4903621ae3))
+- add test component route for development ([f2a397a](https://github.com/giftdavid101/appointment-schedular/commitsf2a397ae557a7818a8a975e2397677901e34bc70))
+- create appointment form ([d5d58c1](https://github.com/giftdavid101/appointment-schedular/commitsd5d58c1e8d93326709eae89ff2b552beb77c3858))
+- create patient overview form ([59b363e](https://github.com/giftdavid101/appointment-schedular/commits59b363e3bcae86f8e0c2ff041c9dd8ff98dbdf16))
+- update standard version config to show chore and styling changes ([1ce6824](https://github.com/giftdavid101/appointment-schedular/commits1ce68242ffd8391ac01d50cb3f2d4c0336db7c04))
+
+### Bug Fixes
+
+- export type colorT from typography ([65f4bc6](https://github.com/giftdavid101/appointment-schedular/commits65f4bc6c358dc79b2af582018aa20585463d06ca))
+- input min-width was too large ([8a73859](https://github.com/giftdavid101/appointment-schedular/commits8a73859e088abdeaa4a2bc25ad78b13b8f388b25))
+- remove components from managePatients to testComponents ([ebafe4f](https://github.com/giftdavid101/appointment-schedular/commitsebafe4fe606170d48e378729735f6dd51f57e7e4))
+- Style-Add style to dashboard. ([593e885](https://github.com/giftdavid101/appointment-schedular/commits593e885f1d6f684583d4ec831c0352d7e8632374))
+
+### Chore
+
+- create doctorsSnippet component. ([6a9a285](https://github.com/giftdavid101/appointment-schedular/commits6a9a2859c2b24775f369db7a70b306d0107dd544))
+- create input component ([e872423](https://github.com/giftdavid101/appointment-schedular/commitse872423c541ddf378538bedc8c2b0d63e8d7b28e))
+
+### Styling
+
+- add more colors to variable ([ef47b77](https://github.com/giftdavid101/appointment-schedular/commitsef47b77267b97eb6fb33c5ecd5d7640add2cb739))
+- fix image radius. ([3a6a18d](https://github.com/giftdavid101/appointment-schedular/commits3a6a18df643cc1b289ab88332ebf00440b00c4b1))
+- general css rules for modal ([7719f1f](https://github.com/giftdavid101/appointment-schedular/commits7719f1f0c06bc13d6098b78ae0c50a9b9ba85330))
+
 ## [1.1.0](https://github.com/giftdavid101/appointment-schedular/compare/v1.0.0...v1.1.0) (2021-10-03)
 
 ### Features